+++ 3_project/src/SW/qna_contoller/DeleteControllor.java
@@ -1,4 +1,4 @@
-package SW_controller;
+package SW.qna_contoller;
 
 import java.io.IOException;
 
@@ -9,38 +9,35 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
 import SW_dao.QnAlistDao;
-import SW_vo.QnAvo;
 
-@WebServlet("/SW_pro/QnAwrite")
-public class QnAwriteController extends HttpServlet {
+@WebServlet("/SW_pro/Delete")
+public class DeleteControllor extends HttpServlet {
 	@Override
 	protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
-		req.setAttribute("top", "/pro/header.jsp");
-		req.setAttribute("main",	"/SW_pro/QnAwrite.jsp");
-		req.setAttribute("bottom", "/pro/footer.jsp");
+		int qanum=Integer.parseInt(req.getParameter("qanum"));
+		req.setAttribute("qanum", qanum);
+		req.setAttribute("top","/pro/header.jsp");
+		req.setAttribute("main","/SW_pro/delete.jsp");
+		req.setAttribute("bottom","/pro/footer.jsp");
 		req.getRequestDispatcher("/pro/product.jsp").forward(req, resp);
 	}
-	
 	@Override
 	protected void doPost(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
 		req.setCharacterEncoding("utf-8");
-		String qacontent=req.getParameter("qacontent");
-		String qaname=req.getParameter("qaname");
+		int qanum=Integer.parseInt(req.getParameter("qanum"));
 		String qapwd=req.getParameter("qapwd");
-		
-		QnAvo vo=new QnAvo(0, qacontent, qaname, qapwd, null, 0, null);
+		System.out.println(qanum);
 		QnAlistDao dao=QnAlistDao.getInstance();
-		int n=dao.insert(vo);
-		if(n>0){
-			req.setAttribute("msg", "success");
+		int n=dao.delete(qanum, qapwd);
+		if(n>0) {
+			req.setAttribute("msg","success");
 		}else {
-			req.setAttribute("msg", "fail");
+			req.setAttribute("msg","fail");
 		}
 		req.setAttribute("top", "/pro/header.jsp");
-		req.setAttribute("main","/SW_pro/result.jsp");
+		req.setAttribute("content","/SW_pro/result.jsp");
 		req.setAttribute("bottom", "/pro/footer.jsp");
 		req.getRequestDispatcher("/pro/product.jsp").forward(req, resp);
 	}
 }
-
