+++ 3_project/src/SW/qna_contoller/QnAlistController.java
@@ -1,4 +1,4 @@
-package SW_controller;
+package SW.qna_contoller;
 
 import java.io.IOException;
 import java.util.ArrayList;
